fix(movie-details): delete state for details and add details page
