+++ movie/src/components/Header/Header/Header.tsx
@@ -1,67 +1,23 @@
-import React, { Dispatch } from "react";
+import React from "react";
 import styled from "styled-components";
-import { Box, createStyles, Grid, makeStyles } from "@material-ui/core";
-import { useDispatch, useSelector } from "react-redux";
+import { Box, Grid } from "@material-ui/core";
 import AddMovieButton from "../AddMovieButton";
 import SearchMovie from "../SearchMovie";
 import Logo from "../Logo";
 import colors from "../../../variables/colors";
-import MovieDetails from "../MovieDetails";
-import searchImage from "../../../img/search-solid.svg";
-import { selectIsShowMovieDetails } from "../../../state/reducers/selectors";
-import { setIsShowMovieDetails } from "../../../state/reducers/actionCreators";
-import { isShowMovieDetailsActions } from "../../../state/actions/isShowMovieDetailsAction";
-
-const useStyles = makeStyles(() =>
-  createStyles({
-    searchImage: {
-      height: "25px",
-      width: "25px",
-      cursor: "pointer",
-    },
-  }),
-);
 
 const Header: React.FC = () => {
-  const isShowMovieDetails = useSelector(selectIsShowMovieDetails);
-  const classes = useStyles();
-  const isShowMovieDispatch = useDispatch<
-    Dispatch<isShowMovieDetailsActions>
-  >();
-
-  const handlerSearch = () => {
-    isShowMovieDispatch(setIsShowMovieDetails(false));
-  };
-
   return (
-    <>
-      {!isShowMovieDetails ? (
-        <HeaderColored>
-          <Box justifyContent="space-between" display="flex">
-            <Logo />
-            <AddMovieButton />
-          </Box>
-          <SearchPanel>
-            <Headline>Find your movie</Headline>
-            <SearchMovie />
-          </SearchPanel>
-        </HeaderColored>
-      ) : (
-        <MovieInfoColored>
-          <Box justifyContent="space-between" display="flex">
-            <Logo />
-            <img
-              src={searchImage}
-              alt="search_image"
-              className={classes.searchImage}
-              onClick={handlerSearch}
-              role="presentation"
-            />
-          </Box>
-          <MovieDetails />
-        </MovieInfoColored>
-      )}
-    </>
+    <HeaderColored>
+      <Box justifyContent="space-between" display="flex">
+        <Logo />
+        <AddMovieButton />
+      </Box>
+      <SearchPanel>
+        <Headline>Find your movie</Headline>
+        <SearchMovie />
+      </SearchPanel>
+    </HeaderColored>
   );
 };
 
@@ -76,14 +32,6 @@ const SearchPanel = styled(Grid)`
 `;
 
 export const HeaderColored = styled(Grid)`
-  width: 100%;
-  background-color: ${colors.darkViolet};
-  padding: 20px 15% 100px;
-  box-sizing: border-box;
-  border-bottom: 10px solid ${colors.lightGrey};
-`;
-
-const MovieInfoColored = styled(Grid)`
   width: 100%;
   background-color: ${colors.darkViolet};
   padding: 20px 15% 100px;
