+++ movie/src/pages/Details/index.ts
@@ -1,0 +1,3 @@
+import DetailsPage from "./Details";
+
+export default DetailsPage;
