+++ movie/src/pages/Home/Home.tsx
@@ -1,0 +1,16 @@
+import React from "react";
+import Footer from "../../components/Footer";
+import Header from "../../components/Header";
+import Main from "../../components/Main";
+
+const HomePage: React.FC = () => {
+  return (
+    <>
+      <Header />
+      <Main />
+      <Footer />
+    </>
+  );
+};
+
+export default HomePage;
