+++ movie/src/pages/Home/index.ts
@@ -1,0 +1,3 @@
+import HomePage from "./Home";
+
+export default HomePage;
