--- movie/src/state/actions/isShowMovieDetailsAction.ts
@@ -1,5 +1,0 @@
-export interface IisShowMovieDetailsAction {
-  readonly type: string;
-  isShowMovieDetails: boolean;
-}
-export type isShowMovieDetailsActions = IisShowMovieDetailsAction;
