+++ movie/src/state/reducers/selectors.ts
@@ -3,7 +3,6 @@ import { AppState } from "./rootRedicer";
 
 const selectMovieSelector = (state: AppState) => state.movie;
 const selectMovies = (state: AppState) => state.allMovies;
-const selectMovieDetails = (state: AppState) => state.isShowMovieDetails;
 
 export const selectMovie = createSelector(
   selectMovieSelector,
@@ -13,9 +12,4 @@ export const selectMovie = createSelector(
 export const selectAllMovies = createSelector(
   selectMovies,
   movieData => movieData,
-);
-
-export const selectIsShowMovieDetails = createSelector(
-  selectMovieDetails,
-  isShow => isShow,
 );
